refactor: extract store jobs
